+++ src/formatting.js
@@ -1,0 +1,9 @@
+import baseConfig from './baseConfig.js';
+import jsConfig from './formatting/jsConfig.js';
+import tsConfig from './formatting/tsConfig.js';
+
+export default [
+	...baseConfig,
+	jsConfig,
+	tsConfig
+];