+++ src/globPatterns.js
@@ -1,0 +1,4 @@
+export const allFiles = '**/*.{js,cjs,mjs,ts,cts,mts}';
+export const jsFiles = '**/*.{js,cjs,mjs}';
+export const tsFiles = '**/*.{ts,cts,mts}';
+