+++ src/index.js
@@ -1,0 +1,11 @@
+import baseConfig from './baseConfig.js';
+import jsConfig from './jsConfig.js';
+import testsConfig from './testsConfig.js';
+import tsConfig from './tsConfig.js';
+
+export default [
+	...baseConfig,
+	jsConfig,
+	tsConfig,
+	testsConfig
+];